Optimize Invoke-ScriptBlock
